chore: rename and move app-bar-render
