+++ src/pages/ProductView.js
@@ -2,7 +2,7 @@ import React from 'react'
 import CircularProgress from '@material-ui/core/CircularProgress'
 import qs from 'querystring'
 import relay from '../createRelay'
-import CartRootRender from './CartRootRender'
+import AppBarRender from './common/AppBarRender'
 import { Product } from './containers'
 import { putProductMutation } from './mutations'
 import { makeStyles } from '@material-ui/styles'
@@ -46,7 +46,7 @@ const ProductView = (props: Props) => {
 
   return (
     <div>
-      <CartRootRender>
+      <AppBarRender>
         {({ refetchCart, cart }) => (
           <div className={classes.root}>
             <QueryRenderer
@@ -94,7 +94,7 @@ const ProductView = (props: Props) => {
             />
           </div>
         )}
-      </CartRootRender>
+      </AppBarRender>
     </div>
   )
 }
