+++ src/pages/containers/index.js
@@ -1,4 +1,4 @@
-export { default as CartRoot } from './CartRoot'
+export { default as AppBar } from './AppBar'
 export { default as Product } from './Product'
 export { default as ProductList } from './ProductList'
 export { default as ProductCartList } from './ProductCartList'
