+++ src/pages/index.js
@@ -1,4 +1,4 @@
-export { default as Home } from './Home.js'
+export { default as HomeView } from './HomeView.js'
 export { default as CartView } from './CartView.js'
 export { default as CheckoutView } from './CheckoutView.js'
 export { default as ProductView } from './ProductView.js'
